bitbake: parse: Conditional support
Teaches bitbake to understand conditionals when parsing recipes. Conditionals
take the form of:

if <python expression>:
	<statement>
	<statement>

If the python expression evaluates to a true value (using the same logic to
determine truth values as python), the statements in the block are evaluated.

The rules for what statements go in a conditional block are based on
indentation: The first statment after the conditional sets the indentation,
which must be at least one whitespace charater more indented than the
conditional itself. All other statements must be of the same indentation to be
in that block. Once a statement with another indentation level is encountered,
the block ends. This is designed to emulate the python block rules.

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>