icecc-create-env: Archive directory
Taring up the toolchain is now done by adding the entire working
directory, instead of listing all the files individually. This is done
because the list of files may contain ".." entries, which tar does not
like and strips out, resulting in bad archives. This should result in an
identical archive to what was previously generated.

In addition, symbolic links are no longer dereferenced when creating the
archive, as they are purposely included to provide alternate names for
files

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>