icecc.bbclass: Skip canadian-cross compiles
icecc.bbclass will no longer attempt to distribute cross-canadian
compiles. While technically possible to generate a toolchain that runs
on the build system and generates executables for the host system, this
is not the normal way that icecc operates and there are so few of these
recipes that it is probably not worth maintaining.

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>